move processing into /backend
