+++ backend/processing/downloader.py
@@ -7,7 +7,7 @@ import requests
 
 from backend.layers.business.business_interface import BusinessLogicInterface
 from backend.layers.common.entities import DatasetStatusKey, DatasetUploadStatus, DatasetVersionId
-from backend.layers.processing.exceptions import UploadFailed
+from backend.processing.exceptions import UploadFailed
 
 logger = logging.getLogger(__name__)
