+++ backend/processing/process_cxg.py
@@ -5,11 +5,11 @@ from backend.layers.common.entities import (
     DatasetStatusKey,
     DatasetVersionId,
 )
-from backend.layers.processing.h5ad_data_file import H5ADDataFile
-from backend.layers.processing.logger import logit
-from backend.layers.processing.process_logic import ProcessingLogic
 from backend.layers.thirdparty.s3_provider import S3ProviderInterface
 from backend.layers.thirdparty.uri_provider import UriProviderInterface
+from backend.processing.h5ad_data_file import H5ADDataFile
+from backend.processing.logger import logit
+from backend.processing.process_logic import ProcessingLogic
 
 
 class ProcessCxg(ProcessingLogic):
