+++ backend/processing/process_seurat.py
@@ -8,10 +8,10 @@ from backend.layers.common.entities import (
     DatasetStatusKey,
     DatasetVersionId,
 )
-from backend.layers.processing.logger import logit
-from backend.layers.processing.process_logic import ProcessingLogic
 from backend.layers.thirdparty.s3_provider import S3ProviderInterface
 from backend.layers.thirdparty.uri_provider import UriProviderInterface
+from backend.processing.logger import logit
+from backend.processing.process_logic import ProcessingLogic
 
 
 class ProcessSeurat(ProcessingLogic):
