+++ backend/processing/schema_migration.py
@@ -21,9 +21,9 @@ from backend.layers.common.entities import (
     DatasetVersion,
     DatasetVersionId,
 )
-from backend.layers.processing import logger
-from backend.layers.processing.process_logic import ProcessingLogic
 from backend.layers.thirdparty.step_function_provider import StepFunctionProvider, sfn_name_generator
+from backend.processing import logger
+from backend.processing.process_logic import ProcessingLogic
 
 logger.configure_logging(level=logging.INFO)
