+++ backend/processing/upload_failures/app.py
@@ -6,8 +6,8 @@ from typing import Optional
 from backend.common.utils.aws import delete_many_from_s3
 from backend.common.utils.result_notification import aws_batch_job_url_fmt_str, aws_sfn_url_fmt_str, notify_slack
 from backend.layers.common.entities import DatasetProcessingStatus, DatasetStatusKey, DatasetVersionId
-from backend.layers.processing import logger
 from backend.portal.api.providers import get_business_logic
+from backend.processing import logger
 
 logger.configure_logging(level=logging.INFO)
