+++ tests/unit/processing/test_processing.py
@@ -7,10 +7,10 @@ from backend.layers.common.entities import (
     DatasetUploadStatus,
     DatasetValidationStatus,
 )
-from backend.layers.processing.process import ProcessMain
-from backend.layers.processing.process_cxg import ProcessCxg
-from backend.layers.processing.process_download_validate import ProcessDownloadValidate
-from backend.layers.processing.process_seurat import ProcessSeurat
+from backend.processing.process import ProcessMain
+from backend.processing.process_cxg import ProcessCxg
+from backend.processing.process_download_validate import ProcessDownloadValidate
+from backend.processing.process_seurat import ProcessSeurat
 from tests.unit.processing.base_processing_test import BaseProcessingTest
